Mating via Sexual Selection

Example of alternative function to govern mating. Fertile females of each phenotype have distinct preferences regarding phenotype of a prospective male mate. The user must set the value of prefs.

sexualSelection(individuals, number_of_couples, prefs = matrix(1, nrow = 3,
  ncol = 3))

Arguments

individuals
Data frame containing id, sex, warner-status, mother id and father id. Value provided.

number_of_couples
Number of couples to form in the current generation. Value provided.

prefs
Numerical matrix with three rows and three columns. First row gives preferences of female of with 0 altrusits alleles, etc. Thus the row (1, 5, 10) indicates that female is ten five more likely to mate with a given 1-allele male than with a given 0-allele male, and ten times more likely to mate with a 2-allele male than a 0-allele male.

Value

A list with two named elements:

females

character vector of id numbers of females

males

character vector of id numbers of females

The two vectors are of the same length. Corresponding elements represent a couple.

Examples

# NOT RUN {
pop <- simulate(sim_gens = 200,
                mating_behavior = list(
                  fn = sexualSelection,
                  args = list(
                    prefs = matrix(
                      c(1, 5, 10, 1, 1, 1, 10, 5, 1),
                      nrow = 3,
                      ncol = 3))))
# }