Simulate Spread of Gene for Altruism

Simulate Spread of Gene for Altruism

simulate(average_litter_size = 5, initial_males = 100,
  initial_alt_males = 10, initial_females = 100, initial_alt_females = 10,
  birth_rate_natural = 0.05, death_rate_natural = 0, prob_attack = 0.2,
  warner_death_prob = 0.4, nonwarner_death_prob = 0.2,
  hider_death_prob = 0, sim_gens = 2, capacity = 2000,
  mating_behavior = NULL, attack_behavior = NULL, graph = TRUE)

Arguments

average_litter_size
Mean size of a litter.

initial_males
Number of males in initial population.

initial_alt_males
Number of males in initial population with two alleles for altruism (all other initial males have 0 such alleles).

initial_females
Number of females in initial population.

initial_alt_females
Number of females in initial population with two alleles for altruism (all other initial females have 0 such alleles).

birth_rate_natural
Birth rate for the population.

death_rate_natural
Death rate for a size-zero population. Rises linearly to birth_rate_natrual as populaiion approaches capacity.

prob_attack
The probability of a predator attack in any given generation.

warner_death_prob
Probability that an individual who warns others during an attack will be killed.

nonwarner_death_prob
Probability of death for an individual who does not warn others but who was not forewarned by a warner.

hider_death_prob
Probability of death during an attack for an individual who accepts a warning.

sim_gens
Number of generations to simulate.

capacity
Carrying capacity of the population.

mating_behavior
Custom function to govern how eligible partners form couples.

attack_behavior
Custom function to govern behavior of warners when population is under attack.

graph
If TRUE, provides a graph of the total per-capita warner alleles in population over the generations.

Value

A data frame with information on the population at each generation. Each row describes the population at the end of a single birth-death-attack cycle. Variables are:

populationSize

total population

males

total number of males

males0

number of males with no alleles for altruism

males1

number of males with one allele for altruism

males2

number of males with two alleles for altruism

females

total number of females

females0

number of females with no alleles for altruism

females1

number of females with one allele for altruism

females2

number of females with two alleles for altruism

Note

For details on mating_behavior and attack_behavior consult the package documentation.

Examples

# NOT RUN {
# use defaults, get a graph:
pop <- simulate(sim_gens = 400)
# attacks are infrequent, and it's dangerous to warn:
pop <- simulate(sim_gens = 200,
                warner_death_prob = 0.8,
                attack_prob = 0.05)
# use an alternative mating function exported by package:
pop <- simulate(sim_gens = 200,
                warner_death_prob = 0.8,
                mating_behavior = list(
                  fn = sexualSelection,
                  args = list(
                    matrix(
                      c(1, 5, 10, 1, 1, 1, 10, 5, 1),
                      nrow = 3,
                      ncol = 3))))
# }